requirements.txt
asgiref==3.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.7
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.23
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
